feat: Translator for external deps, minor improvement to AssetOut (#41)
* feat: move translator to config, add support for kinds label and specify group through translator

The SQLMeshDagsterTranslator has been simplified to return an AssetKey, the name of
the AssetKey or a string equivalent, which deprecates utils.sqlmesh_model_name_to_key
and utils.key_to_sqlmesh_model_name.

The kinds label is now also added, to show technology labels on the UI, only when
Dagster's version allows for it.

Groups can now be specified through the translator, instead of having a fixed method.

* chore: merge source for PR

* fix: rename translator parameters and set defaults

* fix: remove translator from controller and use helper for internal ref

* chore: remove unused refs to translator

* fix: make translator required param for controller

* chore: formatting

* chore: update internal module refs to be absolute

* fix: undo singleton for translator